refactor(generator): extract ContractAndDtoChecks
